+++ src/Commands/AutoLoad.h
@@ -1,8 +1,11 @@
 #pragma once
 
 #include "Commands.h"
+#include <BuildingClass.h>
+#include <TechnoClass.h>
+#include "Ext/Techno/Body.h"
+#include <unordered_map>
 
-// Select next idle harvester
 class AutoLoadCommandClass : public CommandClass
 {
 public:
@@ -12,4 +15,8 @@ public:
 	virtual const wchar_t *GetUICategory() const override;
 	virtual const wchar_t *GetUIDescription() const override;
 	virtual void Execute(WWKey eInput) const override;
+
+	static int GetBuildingPassengerBudget(BuildingClass* pBuilding);
+	static bool CanBeBuildingPassenger(TechnoClass* pPassenger);
+	static std::set<TechnoClass*> SpreadPassengersToTransports(std::vector<TechnoClass*>& passengers, std::vector<std::pair<TechnoClass*, int>>& transports);
 };