+++ src/Commands/Commands.cpp
@@ -11,6 +11,7 @@
 #include "ToggleDesignatorRange.h"
 #include "SaveVariablesToFile.h"
 #include "AutoLoad.h"
+#include "AutoGarrison.h"
 
 DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 {
@@ -21,6 +22,7 @@ DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 	MakeCommand<ToggleDigitalDisplayCommandClass>();
 	MakeCommand<ToggleDesignatorRangeCommandClass>();
 	MakeCommand<AutoLoadCommandClass>();
+	MakeCommand<AutoGarrisonCommandClass>();
 
 	if (Phobos::Config::DevelopmentCommands)
 	{